requirements.txt
bs4==0.0.2
cohere==5.13.11
configobj==5.0.9
configparser==7.1.0
docling==2.23.1
Flask==3.1.0
httplib2==0.22.0
httptools==0.6.4
langchain-chroma==0.1.4
langchain-community==0.3.5
langchain-openai==0.2.5
langchainhub==0.1.21
langgraph==0.2.45
markitdown==0.0.1a1
multiprocess==0.70.17
nipype==1.9.0
nltk==3.9.1
PyMuPDF==1.24.13
pyxnat==1.6.2
rank-bm25==0.2.2
semchunk==2.2.2
sentence-transformers==3.4.1
streamlit==1.40.2
watchfiles==0.24.0
websockets==13.1
